chore: migrate DOM methods away from IsolatedWorld
