+++ test/src/click.spec.ts
@@ -52,24 +52,6 @@ describe('Page.click', function () {
       })
     ).toBe(42);
   });
-  itFailsFirefox(
-    'should click the button if window.Node is removed',
-    async () => {
-      const {page, server} = getTestState();
-
-      await page.goto(server.PREFIX + '/input/button.html');
-      await page.evaluate(() => {
-        // @ts-expect-error Expected.
-        return delete window.Node;
-      });
-      await page.click('button');
-      expect(
-        await page.evaluate(() => {
-          return (globalThis as any).result;
-        })
-      ).toBe('Clicked');
-    }
-  );
   // @see https://github.com/puppeteer/puppeteer/issues/4281
   it('should click on a span with an inline element inside', async () => {
     const {page} = getTestState();
